# artifacts/api-server/src/routes/bulletins.ts
import { Router } from "express";
import { readFile } from "node:fs/promises";
import {
  GetBulletinStatusResponse,
  GetCurrentBulletinResponse,
  GetBulletinAudioParams,
  ListBulletinHistoryResponse,
  ListBulletinTransmissionRoomsResponse,
  ListBulletinTransmissionsResponse,
  TransmitBulletinParams,
  TransmitBulletinBody,
  TransmitBulletinResponse,
  GetBulletinScheduleResponse,
  UpdateBulletinScheduleBody,
} from "@workspace/api-zod";
import { requireAdmin } from "../lib/adminMiddleware";
import { logger } from "../lib/logger";
import {
  AEMET_SOURCE_URL,
} from "../lib/bulletins/aemet";
import {
  getAudioPath,
  listBulletins,
  type StoredBulletin,
  listBulletinTransmissions,
} from "../lib/bulletins/storage";
import { transmitBulletin } from "../lib/bulletins/transmission";
import { generateBulletin } from "../lib/bulletins/generation";
import { bulletinScheduler, requireSchedulerReady } from "../lib/bulletins/scheduler";
import { roomManager } from "../eqso/room-manager";

const router = Router();
const IDENTITY = "eQSO Sierra Noroeste";
const GEOGRAPHIC_FOCUS = "Los Barrancos, Valdemorillo, Sierra Noroeste de Madrid";

function toResponse(bulletin: StoredBulletin): StoredBulletin & { audioUrl: string } {
  return { ...bulletin, audioUrl: `/api/admin/bulletins/${bulletin.id}/audio` };
}

router.use(requireAdmin);

router.get("/bulletins/schedule", (_req, res) => {
  try {
    requireSchedulerReady();
    res.json(GetBulletinScheduleResponse.parse(bulletinScheduler.status()));
  } catch (error) {
    res.status(503).json({ error: error instanceof Error ? error.message : "Programador no disponible" });
  }
});

router.put("/bulletins/schedule", async (req, res) => {
  const input = UpdateBulletinScheduleBody.safeParse(req.body);
  if (!input.success) {
    res.status(400).json({ error: "Configuración inválida: revise intervalos, sala y opciones" });
    return;
  }
  if (input.data.transmissionEnabled && (!input.data.confirmed || !roomManager.getRooms().includes(input.data.room))) {
    res.status(400).json({ error: "Seleccione una sala válida y confirme las emisiones RF automáticas" });
    return;
  }
  try {
    requireSchedulerReady();
    const result = await bulletinScheduler.configure(input.data);
    logger.info({ config: result.config, requestedBy: req.session?.callsign }, "Programación de boletines guardada");
    res.json(GetBulletinScheduleResponse.parse(result));
  } catch (error) {
    req.log.error({ err: error }, "No se pudo guardar la programación");
    res.status(503).json({ error: error instanceof Error ? error.message : "No se pudo guardar" });
  }
});

router.get("/bulletins/status", async (req, res): Promise<void> => {
  try {
    const bulletins = await listBulletins();
    const current = bulletins[0];
    res.json(GetBulletinStatusResponse.parse({
      identity: IDENTITY,
      geographicFocus: GEOGRAPHIC_FOCUS,
      sourceUrl: AEMET_SOURCE_URL,
      historyCount: bulletins.length,
      hasCurrent: Boolean(current),
      currentId: current?.id ?? null,
      currentGeneratedAt: current?.generatedAt ?? null,
    }));
  } catch (error) {
    req.log.error({ err: error }, "No se pudo leer el estado de boletines");
    res.status(500).json({ error: "No se pudo leer el estado de boletines" });
  }
});

router.get("/bulletins/current", async (req, res): Promise<void> => {
  try {
    const [current] = await listBulletins();
    if (!current) {
      res.status(404).json({ error: "Todavía no hay un boletín generado" });
      return;
    }
    res.json(GetCurrentBulletinResponse.parse(toResponse(current)));
  } catch (error) {
    req.log.error({ err: error }, "No se pudo leer el boletín actual");
    res.status(500).json({ error: "No se pudo leer el boletín actual" });
  }
});

router.get("/bulletins/history", async (req, res): Promise<void> => {
  try {
    const bulletins = await listBulletins();
    res.json(ListBulletinHistoryResponse.parse(bulletins.map(toResponse)));
  } catch (error) {
    req.log.error({ err: error }, "No se pudo leer el historial de boletines");
    res.status(500).json({ error: "No se pudo leer el historial de boletines" });
  }
});

router.get("/bulletins/rooms", async (_req, res): Promise<void> => {
  res.json(ListBulletinTransmissionRoomsResponse.parse({ rooms: roomManager.getRooms() }));
});

router.get("/bulletins/transmissions", async (_req, res): Promise<void> => {
  try {
    res.json(ListBulletinTransmissionsResponse.parse(await listBulletinTransmissions()));
  } catch (error) {
    logger.error({ err: error }, "No se pudo leer el historial de transmisiones");
    res.status(500).json({ error: "No se pudo leer el historial de transmisiones" });
  }
});

router.post("/bulletins/generate", async (req, res): Promise<void> => {
  try {
    const bulletin = await generateBulletin();
    res.status(201).json(GetCurrentBulletinResponse.parse(toResponse(bulletin)));
  } catch (error) {
    req.log.error({ err: error }, "No se pudo generar el boletín meteorológico");
    res.status(502).json({
      error: error instanceof Error ? error.message : "No se pudo generar el boletín meteorológico",
    });
  }
});

router.post("/bulletins/:id/transmit", async (req, res): Promise<void> => {
  const params = TransmitBulletinParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: "Identificador de boletín inválido" });
    return;
  }
  const body = TransmitBulletinBody.safeParse(req.body);
  if (!body.success || body.data.confirmed !== true) {
    res.status(400).json({ error: "La confirmación debe ser true" });
    return;
  }
  const room = body.data.room;
  if (!roomManager.getRooms().includes(room)) {
    res.status(400).json({ error: "Sala no válida" });
    return;
  }
  try {
    const bulletin = (await listBulletins()).find((item) => item.id === params.data.id);
    if (!bulletin) {
      res.status(404).json({ error: "Boletín no encontrado" });
      return;
    }
    const result = await transmitBulletin(bulletin, room, req.session?.callsign ?? null);
    const payload = TransmitBulletinResponse.parse(result.record);
    res.status(result.rejected ? 409 : result.record.status === "failed" ? 502 : 200).json(payload);
  } catch (error) {
    req.log.error({ err: error, bulletinId: params.data.id, room }, "No se pudo iniciar la transmisión del boletín");
    res.status(502).json({ error: error instanceof Error ? error.message : "No se pudo transmitir el boletín" });
  }
});

router.get("/bulletins/:id/audio", async (req, res): Promise<void> => {
  const params = GetBulletinAudioParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: "Identificador de boletín inválido" });
    return;
  }
  try {
    const bulletin = (await listBulletins()).find((item) => item.id === params.data.id);
    if (!bulletin) {
      res.status(404).json({ error: "Boletín no encontrado" });
      return;
    }
    const audio = await readFile(getAudioPath(bulletin));
    res.type(bulletin.audioMimeType);
    res.send(audio);
  } catch (error) {
    req.log.error({ err: error }, "No se pudo leer el audio del boletín");
    res.status(404).json({ error: "Audio del boletín no encontrado" });
  }
});

export default router;