# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface BulletinScheduleInput {
  generationEnabled: boolean;
  transmissionEnabled: boolean;
  /**
   * @minimum 15
   * @maximum 10080
   */
  generationIntervalMinutes: number;
  /**
   * @minimum 5
   * @maximum 10080
   */
  transmissionIntervalMinutes: number;
  /**
   * @minimum 15
   * @maximum 10080
   */
  maxAgeMinutes: number;
  /** @maxLength 100 */
  room: string;
  confirmed?: boolean;
}

export type BulletinScheduleEventStatus =
  (typeof BulletinScheduleEventStatus)[keyof typeof BulletinScheduleEventStatus];

export const BulletinScheduleEventStatus = {
  completed: "completed",
  failed: "failed",
  skipped: "skipped",
} as const;

export interface BulletinScheduleEvent {
  at: string;
  status: BulletinScheduleEventStatus;
  message: string;
}

export interface BulletinScheduleStatus {
  config: BulletinScheduleInput;
  /** @nullable */
  nextGenerationAt: string | null;
  /** @nullable */
  nextTransmissionAt: string | null;
  generationRunning: boolean;
  transmissionRunning: boolean;
  lastGeneration: BulletinScheduleEvent | null;
  lastTransmission: BulletinScheduleEvent | null;
}

export interface RelayTokenInfo {
  id: number;
  callsign: string;
  label: string;
  legacyDisabled: boolean;
  createdAt: string;
  /** @nullable */
  lastUsedAt: string | null;
  /** @nullable */
  revokedAt: string | null;
}

export interface RelayTokenCreated {
  id: number;
  callsign: string;
  label: string;
  token: string;
  legacyDisabled: boolean;
  createdAt: string;
  /** @nullable */
  lastUsedAt: string | null;
  /** @nullable */
  revokedAt: string | null;
}

export interface RelayTokenInput {
  /**
   * @minLength 4
   * @maxLength 30
   */
  callsign: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  label: string;
}

export interface RelayTokenRevoked {
  ok: boolean;
}

export interface HealthStatus {
  status: string;
}

export interface WeatherBulletin {
  id: string;
  identity: string;
  geographicFocus: string;
  municipality: string;
  forecastText: string;
  sourceUrl: string;
  sourceAttribution: string;
  /** @nullable */
  sourcePublishedAt: string | null;
  sourceRetrievedAt: string;
  generatedAt: string;
  audioFileName: string;
  audioMimeType: string;
  audioUrl: string;
  forecastDates: string[];
}

export interface BulletinStatus {
  identity: string;
  geographicFocus: string;
  sourceUrl: string;
  historyCount: number;
  hasCurrent: boolean;
  /** @nullable */
  currentId: string | null;
  /** @nullable */
  currentGeneratedAt: string | null;
}

export interface BulletinTransmissionRooms {
  rooms: string[];
}

export interface BulletinTransmissionInput {
  /** @minLength 1 */
  room: string;
  confirmed: true;
}

export type BulletinTransmissionStatus =
  (typeof BulletinTransmissionStatus)[keyof typeof BulletinTransmissionStatus];

export const BulletinTransmissionStatus = {
  completed: "completed",
  failed: "failed",
  rejected: "rejected",
} as const;

export interface BulletinTransmission {
  id: string;
  bulletinId: string;
  bulletinGeneratedAt: string;
  room: string;
  requestedAt: string;
  /** @nullable */
  startedAt: string | null;
  /** @nullable */
  completedAt: string | null;
  status: BulletinTransmissionStatus;
  /** @nullable */
  error: string | null;
  packetCount: number;
  durationMs: number;
  /** @nullable */
  requestedBy: string | null;
}