# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/bulletins/{id}/transmit:
    post:
      operationId: transmitBulletin
      tags: [bulletins]
      summary: Manually transmit an existing bulletin to one room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulletinTransmissionInput"
      responses:
        "200":
          description: Transmission completed or failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulletinTransmission"
        "400":
          description: Invalid bulletin, room, or confirmation
        "404":
          description: Bulletin not found
        "409":
          description: Room is already transmitting
        "502":
          description: Audio conversion or transmission failure
  /admin/bulletins/generate:
    post:
      operationId: generateBulletin
      tags: [bulletins]
      summary: Generate a weather bulletin manually
      responses:
        "201":
          description: Generated weather bulletin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeatherBulletin"
        "502":
          description: AEMET or audio integration failure
  /admin/bulletins/{id}/audio:
    get:
      operationId: getBulletinAudio
      tags: [bulletins]
      summary: Download generated bulletin audio
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Bulletin audio in WAV format
          content:
            audio/wav:
              schema:
                type: string
                format: binary
        "404":
          description: Audio or bulletin not found
components:
  schemas:
...
    WeatherBulletin:
      type: object
      properties:
        id:
          type: string
          format: uuid
        identity:
          type: string
        geographicFocus:
          type: string
        municipality:
          type: string
        forecastText:
          type: string
        sourceUrl:
          type: string
          format: uri
        sourceAttribution:
          type: string
        sourcePublishedAt:
          type: ["string", "null"]
        sourceRetrievedAt:
          type: string
          format: date-time
        generatedAt:
          type: string
          format: date-time
        audioFileName:
          type: string
        audioMimeType:
          type: string
        audioUrl:
          type: string
        forecastDates:
          type: array
          items:
            type: string
            format: date
      required:
        - id
        - identity
        - geographicFocus
        - municipality
        - forecastText
        - sourceUrl
        - sourceAttribution
        - sourcePublishedAt
        - sourceRetrievedAt
        - generatedAt
        - audioFileName
        - audioMimeType
        - audioUrl
        - forecastDates
...
    BulletinTransmissionInput:
      type: object
      properties:
        room:
          type: string
          minLength: 1
        confirmed:
          type: boolean
          const: true
      required:
        - room
        - confirmed
    BulletinTransmission:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bulletinId:
          type: string
          format: uuid
        bulletinGeneratedAt:
          type: string
          format: date-time
        room:
          type: string
        requestedAt:
          type: string
          format: date-time
        startedAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [completed, failed, rejected]
        error:
          type: ["string", "null"]
        packetCount:
          type: integer
        durationMs:
          type: integer
        requestedBy:
          type: ["string", "null"]
      required:
        - id
        - bulletinId
        - bulletinGeneratedAt
        - room
        - requestedAt
        - startedAt
        - completedAt
        - status
        - error
        - packetCount
        - durationMs
        - requestedBy